refactor(Comment): add own Comment component
