--- src/components/PostComments.tsx
@@ -1,89 +1,0 @@
-import { Avatar, AvatarImage, AvatarFallback } from "@radix-ui/react-avatar";
-import { UserCircleIcon } from "lucide-react";
-import format from "date-fns/format";
-import { TypographyMuted } from "./Typography/TypographyMuted";
-import { Typography } from "./Typography/Typography";
-import { TypographySmall } from "./Typography/TypographySmall";
-import { Separator } from "~/components/ui/separator";
-import { CommentDropdown } from "./CommentDropdown";
-import { useSession } from "next-auth/react";
-import CommentActions from "./CommentActions";
-
-type PostCommentsProps = {
-  postName: string;
-  comments: {
-    user: {
-      id: string;
-      name: string;
-      email: string | null;
-      emailVerified: Date | null;
-      image: string | null;
-    };
-    id: string;
-    postId: string;
-    userId: string;
-    text: string;
-    createdAt: Date;
-    updatedAt: Date;
-    originalCommentId: string | null;
-  }[];
-};
-
-export const PostComments = ({ comments, postName }: PostCommentsProps) => {
-  const { data } = useSession();
-  const commentElements = comments.map((comment) => {
-    const date = comment.createdAt;
-    const formattedDate =
-      comments &&
-      format(
-        date,
-        "MMM d" +
-          (date.getFullYear() == new Date().getFullYear() ? "" : ", YYYY"),
-      );
-
-    return (
-      <div key={comment.id} className="flex gap-4 py-6 first:pt-0">
-        <div className="flex flex-col items-center gap-2">
-          <Avatar>
-            {comment.user.image ? (
-              <AvatarImage
-                src={comment.user.image}
-                className="h-10 w-10 rounded-full"
-              />
-            ) : (
-              <AvatarFallback>
-                <UserCircleIcon />
-              </AvatarFallback>
-            )}
-          </Avatar>
-          <Separator orientation="vertical" />
-        </div>
-        <div className="flex flex-col gap-6">
-          <div className="flex flex-col gap-2">
-            <div className="flex gap-4">
-              <Typography>{comment.user.name}</Typography>
-              <TypographyMuted>{formattedDate}</TypographyMuted>
-            </div>
-            <TypographySmall>{comment.text}</TypographySmall>
-          </div>
-          <div className="flex gap-4">
-            <CommentActions
-              session={data}
-              commentId={comment.id}
-              postName={postName}
-            />
-          </div>
-        </div>
-        <div className="ml-auto">
-          <CommentDropdown
-            userIsOrig={data?.user.id === comment.userId}
-            userRole={data?.user.role}
-            commentId={comment.id}
-            postName={postName}
-          />
-        </div>
-      </div>
-    );
-  });
-  return <>{commentElements}</>;
-};
